fix: Wrong setup of Cursor rules
After testing BMAD Method in Cursor rules with the existing setup of the rule
I have noticed that it was not actually using the right prompt. That was noticed
before it was not returning the available commands when opening a new chat and
refering to the agent I needed. I did the changes in my project and the fix wa
to change to "Apply Manually" so the rule `@<agent-name>` has to be referenced
from the chat window, similar to what you do with Claude Code.